Move util to src folder
